big security fix

+++ ProgressCRM/src/java/org/progress/crm/controllers/NewsController.java
@@ -11,6 +11,7 @@ import javax.ejb.EJB;
 import javax.ejb.Singleton;
 import org.hibernate.Session;
 import org.progress.crm.dao.DaoFactory;
+import org.progress.crm.exceptions.BadRequestException;
 import org.progress.crm.exceptions.IsNotAuthenticatedException;
 import org.progress.crm.logic.News;
 
@@ -20,7 +21,10 @@ public class NewsController {
     @EJB
     AuthenticationManager authenticationManager;
 
-    public List<News> getNews(Session session) throws SQLException {
+    public List<News> getNews(Session session, String token) throws SQLException, IsNotAuthenticatedException {
+        if (token == null) {
+            throw new IsNotAuthenticatedException();
+        }
         return DaoFactory.getNewsDao().getNews(session);
     }
 
@@ -34,26 +38,29 @@ public class NewsController {
         return true;
     }
 
-    public boolean deleteNewsById(Session session, String token, String id) throws IsNotAuthenticatedException {
+    public boolean deleteNewsById(Session session, String token, String id) throws IsNotAuthenticatedException, BadRequestException {
+        if (id == null) {
+            throw new BadRequestException();
+        }
         if (token == null) {
             throw new IsNotAuthenticatedException();
         }
-        //FIXME add permission
         UUID uuid = UUID.fromString(token);
         int idWorker = authenticationManager.getUserIdByToken(uuid);
         DaoFactory.getNewsDao().removeNewsById(session, idWorker, Integer.valueOf(id));
         return true;
     }
 
-    public boolean editNewsById(Session session, String newsId, String token, String text, String header) throws IsNotAuthenticatedException {
+    public boolean editNewsById(Session session, String newsId, String token, String text, String header) throws IsNotAuthenticatedException, BadRequestException {
+        if (newsId == null) {
+            throw new BadRequestException();
+        }
         if (token == null) {
             throw new IsNotAuthenticatedException();
         }
-        //FIXME add permission
         UUID uuid = UUID.fromString(token);
         int idWorker = authenticationManager.getUserIdByToken(uuid);
         DaoFactory.getNewsDao().editNewsById(session, Integer.valueOf(newsId), idWorker, header, text);
         return true;
     }
-
 }
